Use get_absolute_url on List model to DRY urls in views
Model objects are often associated with a particular URL, we can use
get_absolute_url which returns the canonical URL for in which the object
is displayed.

If an object defines get_absolute_url, the object editing page in Django
admin will have a "View on site" link that will jump directly to
object's public view.

Even if we change the URL structure of object, we don't have to modify
it every place that it occurs. get_absolute_url will do that for us.